fix: scrub and arrow strokes are not working for transition content and delay (#3832)
## Description

fixes #3821 
Fixes the bug related to updating values for `scrub` and `key strokes`. 

@TrySound @kof was there anything changed in the
`CSSValueInputContainer` or something recently. Because, i see the
values are `unit` values when it is scrub and key down. But it's sending
the whole parsed value when it is just a `onChange` update. Faced a
similar issue in #3812 too.

Steps to reproduce
- Add a transition layer.
- Change the values inside the delay and duration using scrub
- Now, try to repeat the same using key up and down arrows.
- The values will reset back to the original values.


## Code Review

- [ ] hi @kof, I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [x] made a self-review

## Before merging

- [ ] tested locally and on preview environment (preview dev login:
5de6)