experimental: add backface-visibility property for transforms (#3802)
## Description

Allow to add `backface-visibility` property in the transforms section.
`hidden` is not possible on the `backface-visibility` property. Because
it takes only `visible` or `hidden` as value. The other way to simulate
hidden is to reset the value to `visible` because that's the default
value for property. But i don't think it's really needed to manage this.

## Steps for reproduction

1. Click on the `+` icon on the transforms section.
2. Select `Backface Visibility`
3. This will add the property and show the value in the transform layer
list.
4. Click on the layer list item and you can change the property value.


## Code Review

- [ ] hi @kof, I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [x] made a self-review
- [x] added inline comments where things may be not obvious (the "why",
not "what")

## Before merging

- [x] tested locally and on preview environment (preview dev login:
5de6)