fix: better sorting of transitions and allow `enter` key to save (#3968)
fixes #3958 

## Description
This PR fixes the `enter` key behaviour for the transition-property
combo box. And we updated the sort mechanism at the time of searching to
show the most related property at the top. So, users don't need to
scroll all the way to see the matched property.

## Steps for reproduction

- Add a transition property layer.
- Now, click on the layer and the UI panel to edit the properties opens
up.
- Type `height` or any other property in the combo box.
- Now, press `Enter`. The action should save the value of the property
and update the layer.
- Next, search for some random property with a possibility for more
results. Eg: `width`.
- `width` or any other matching value that you searched for should
always show up at the top of the list.

## Code Review

- [ ] hi @kof, I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [x] made a self-review
- [x] added inline comments where things may be not obvious (the "why",
not "what")

## Before merging

- [x] tested locally and on preview environment (preview dev login:
5de6)