feat: allow deleting variables used in expressions (#4860)
Ref https://github.com/webstudio-is/webstudio/issues/4768

Before we disabled deleting variables which are used in expressions. Now
we will ask confirmation and unset this variable. This way user will not
be locked if variable is lost and we will show diagnostics (in the
future) where all unset variables are.

<img width="908" alt="Screenshot 2025-02-12 at 15 38 14"
src="https://github.com/user-attachments/assets/ea9eea58-c276-4251-8b95-1214e9380621"
/>