fix: set xml node background to white (#5183)
## Description

fixes #5015 
Added `backgroundColor` to `white` for `XMLNode` so the content is
visible even when the `GlobalRoot` is set to something else.

## Steps for reproduction

- Create a demo project.
- Create a XML with sitemap.xml
- Now, switch back to the home page and set the `backgroudnColor` of the
global root to some random colour.
- Then, switch back to the sitemap page. Now, the contents are not
visible as the background is inherited from the GlobalRoot. This fix
will help in setting to white whenever a xml node is used.

## Code Review
- [ ] hi @kof, I need you to do
  - test it on preview

## Before requesting a review
- [x] made a self-review

## Before merging
- [x] tested locally and on preview environment (preview dev login:
0000)